requirements.txt
streamlit
python-docx
pdfplumber
PyMuPDF
docx2txt
google-generativeai
python-dotenv
